third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-chorded-buttons.html

<!DOCTYPE HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style>
div.box {
  margin: 10px;
  padding: 50px;
  float: left;
}
</style>

<h1>PointerEvent: Verifies that chorded buttons fire appropriate events</h1>

<div id="target" class="box" style="background-color:red"></div>

<div id="log"></div>

<script>
var receivedEvents = [];
var preventDefaultOnPointerDown = false;

function testReceivedEvents(expectedEvents, testLabel) {
  test(function() {
    assert_array_equals(receivedEvents, expectedEvents);
  }, testLabel);
  receivedEvents = [];
}

function init() {
  var targetDiv = document.getElementById("target");
  var targetEvents = ["mousedown", "mouseup", "pointerdown", "pointerup", "mousemove", "pointermove"];

  targetEvents.forEach(function(eventName) {
    targetDiv.addEventListener(eventName, function(event) {
      receivedEvents.push(event.type + " " + event.button + "," + event.buttons);
      if (event.type.startsWith("mouse"))
        event.preventDefault();
      if (event.type == "pointerdown" && preventDefaultOnPointerDown)
        event.preventDefault();
    });
  });
}

function runTests() {
  var rect = document.getElementById("target").getBoundingClientRect();
  eventSender.mouseMoveTo(rect.left + 5, rect.top + 5);
  receivedEvents = [];

  eventSender.mouseDown(0);
  eventSender.mouseDown(1);
  eventSender.mouseUp(1);
  eventSender.mouseUp(0);
  testReceivedEvents([
    "pointerdown 0,1", "mousedown 0,1",
    "pointermove 1,5", "mousedown 1,5",
    "pointermove 1,1", "mouseup 1,1",
    "pointerup 0,0", "mouseup 0,0"
  ], "LM button sequence A");

  eventSender.mouseDown(0);
  eventSender.mouseDown(1);
  eventSender.mouseUp(0);
  eventSender.mouseUp(1);
  testReceivedEvents([
    "pointerdown 0,1", "mousedown 0,1",
    "pointermove 1,5", "mousedown 1,5",
    "pointermove 0,4", "mouseup 0,4",
    "pointerup 1,0", "mouseup 1,0"
  ], "LM button sequence B");

  eventSender.mouseDown(1);
  eventSender.mouseDown(0);
  eventSender.mouseUp(1);
  eventSender.mouseUp(0);
  testReceivedEvents([
    "pointerdown 1,4", "mousedown 1,4",
    "pointermove 0,5", "mousedown 0,5",
    "pointermove 1,1", "mouseup 1,1",
    "pointerup 0,0", "mouseup 0,0"
  ], "LM button sequence C");

  eventSender.mouseDown(0);
  eventSender.mouseDown(2);
  eventSender.mouseDown(1);
  eventSender.mouseUp(0);
  eventSender.mouseUp(1);
  eventSender.mouseUp(2);
  testReceivedEvents([
    "pointerdown 0,1", "mousedown 0,1",
    "pointermove 2,3", "mousedown 2,3",
    "pointermove 1,7", "mousedown 1,7",
    "pointermove 0,6", "mouseup 0,6",
    "pointermove 1,2", "mouseup 1,2",
    "pointerup 2,0", "mouseup 2,0"
  ], "LMR button sequence");

  eventSender.mouseDown(0);
  eventSender.mouseDown(2);
  eventSender.mouseDown(1);
  eventSender.mouseDown(3);
  eventSender.mouseDown(4);
  eventSender.mouseUp(0);
  eventSender.mouseUp(2);
  eventSender.mouseUp(1);
  eventSender.mouseUp(3);
  eventSender.mouseUp(4);
  testReceivedEvents([
    "pointerdown 0,1", "mousedown 0,1",
    "pointermove 2,3", "mousedown 2,3",
    "pointermove 1,7", "mousedown 1,7",
    "pointermove 3,15", "mousedown 3,15",
    "pointermove 4,31", "mousedown 4,31",
    "pointermove 0,30", "mouseup 0,30",
    "pointermove 2,28", "mouseup 2,28",
    "pointermove 1,24", "mouseup 1,24",
    "pointermove 3,16", "mouseup 3,16",
    "pointerup 4,0", "mouseup 4,0",
  ], "LMRBF button sequence");

  eventSender.mouseDown(1);
  eventSender.mouseMoveTo(rect.left + 6, rect.top + 6);
  eventSender.mouseUp(1);
  eventSender.mouseMoveTo(rect.left + 5, rect.top + 5);
  testReceivedEvents([
    "pointerdown 1,4", "mousedown 1,4",
    "pointermove -1,4", "mousemove 0,4",
    "pointerup 1,0", "mouseup 1,0",
    "pointermove -1,0", "mousemove 0,0",
  ], "Move with & without button");

  eventSender.setMouseButtonState(1); // Off-the-page L-up M-down
  eventSender.mouseDown(0);
  eventSender.setMouseButtonState(2, ["leftButton", "middleButton", "rightButton"]); // Off-the-page R-down
  eventSender.mouseUp(1);
  eventSender.mouseUp(2);
  eventSender.mouseUp(0);
  testReceivedEvents([
    "pointermove 0,5", "mousedown 0,5",
    "pointermove 1,3", "mouseup 1,3",
    "pointermove 2,1", "mouseup 2,1",
    "pointerup 0,0", "mouseup 0,0",
  ], "Off-the-page button changes");

  preventDefaultOnPointerDown = true;
  eventSender.mouseDown(0);
  preventDefaultOnPointerDown = false;
  eventSender.mouseDown(1);
  eventSender.mouseUp(1);
  eventSender.mouseUp(0);
  eventSender.mouseDown(0);
  eventSender.mouseUp(0);
  testReceivedEvents([
    "pointerdown 0,1",
    "pointermove 1,5",
    "pointermove 1,1",
    "pointerup 0,0",
    "pointerdown 0,1", "mousedown 0,1",
    "pointerup 0,0", "mouseup 0,0"
  ], "LM button with preventDefault on pointerdown");
}

test(function() {
  if (!window.eventSender)
    assert_true(true, "No eventSender, skipped tests");
  else if (!window.PointerEvent)
    assert_true(true, "No PointerEvent, skipped tests");
  else {
    init();
    runTests();
  }

  done();
}, "PointerEvent: Verifies that chorded buttons fire appropriate events");
</script>
